+++ system/src/Grav/Common/Flex/Types/Users/UserIndex.php
@@ -118,15 +118,15 @@ class UserIndex extends FlexIndex
     }
 
     /**
-     * @param array $entry
+     * @param array $meta
      * @param array $data
      */
-    protected static function updateIndexData(array &$entry, array $data)
+    public static function updateObjectMeta(array &$meta, array $data)
     {
         // Username can also be number and stored as such.
-        $key = (string)($data['username'] ?? $entry['key']);
-        $entry['key'] = mb_strtolower($key);
-        $entry['email'] = isset($data['email']) ? mb_strtolower($data['email']) : null;
+        $key = (string)($data['username'] ?? $meta['key']);
+        $meta['key'] = mb_strtolower($key);
+        $meta['email'] = isset($data['email']) ? mb_strtolower($data['email']) : null;
     }
 
     /**