+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     * Fixed flex objects with integer keys not working [#2863](https://github.com/getgrav/grav/issues/2863)
     * Fixed user avatar creation for new `Flex Users` when using folder storage
     * Fixed `Trying to access array offset on value of type null` PHP 7.4 error in `Plugin.php`
+    * Fixed Gregwar Image library using `.jpeg` for cached images, rather use `.jpg`
 
 # v1.7.0-rc.8
 ## 03/19/2020