+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ## mm/dd/2020
 
 1. [](#new)
-    * Support for `webp` image format in Page Media #1168
+    * Support for `webp` image format in Page Media [#1168](https://github.com/getgrav/grav/issues/1168)
     * Added `Route::getBase()` method
 1. [](#improved)
     * Support symlinks when saving `File`