+++ CHANGELOG.md
@@ -17,6 +17,7 @@
     * Fixed `Flex Pages` with `00.home` page not having ordering set
     * Fixed `Flex Pages` not updating empty content on save [#2890](https://github.com/getgrav/grav/issues/2890)
     * Fixed creating new Flex User with file storage
+    * Fixed saving new `Flex Object` with custom key
 
 # v1.7.0-rc.8
 ## 03/19/2020